In progress: debugging pgvec.py
