+++ langroid/vector_store/__init__.py
@@ -50,12 +50,12 @@ except ImportError:
     pass
 
 try:
-    from . import pgvector
-    from .pgvector import PGVector, PGVectorConfig
+    from . import pgvec
+    from .pgvec import PGVector, PGVectorConfig
 
-    pgvector
+    pgvec
     PGVector
     PGVectorConfig
-    __all__.extend(["pgvector", "PGVector", "PGVectorConfig"])
+    __all__.extend(["pgvec", "PGVector", "PGVectorConfig"])
 except ImportError:
     pass
