+++ langroid/vector_store/base.py
@@ -58,7 +58,7 @@ class VectorStore(ABC):
         from langroid.vector_store.lancedb import LanceDB, LanceDBConfig
         from langroid.vector_store.meilisearch import MeiliSearch, MeiliSearchConfig
         from langroid.vector_store.momento import MomentoVI, MomentoVIConfig
-        from langroid.vector_store.pgvector import PGVector, PGVectorConfig
+        from langroid.vector_store.pgvec import PGVector, PGVectorConfig
         from langroid.vector_store.qdrantdb import QdrantDB, QdrantDBConfig
 
         if isinstance(config, QdrantDBConfig):
